edac: Introduce armv8.2 RAS extentions-based EDAC driver
The armv8.2 architecture mandates presence of RAS extensions, which
enable on-chip error detection. Add a driver to enable that
functionality. This revision provides support for CPU Ln$, but can be
extended to other entities supported by the spec (such as SMMUs or
GIC).

The driver reuses the ACPI structs for both ACPI (where we can simply
copy the binary data) and DT (where we read some properties and
translate it into what ACPI would give us where it makes sense).

This is somewhat based on previous attempts by Qualcomm and Ampere
Computing.