[very very WIP] edac: Introduce ARM RAS-based EDAC driver
The armv8.2 architecture mandates presence of RAS extensions, which
enable on-chip error detection. Add a driver to enable that
functionality. This revision provides support for CPU Ln$, but can be
extended to other entities supported by the spec (such as SMMUs or
GIC).

This is based on previous efforts by <TODO>.